% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read.shimmer}
\alias{read.shimmer}
\title{Read data from a Shimmer device.}
\usage{
read.shimmer(filename)
}
\arguments{
\item{filename}{The filename with the Shimmer data.}
}
\value{
The data as a recording structure (list).
}
\description{
This function reads data from a Shimmer device (in the CSV format).
}

