% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shimmer.R
\docType{package}
\name{shimmer}
\alias{shimmer}
\alias{shimmer-package}
\title{Read Physiologic Signal Data Recorded with a Shimmer device.}
\description{
The shimmer package contains functions for reading data from a Shimmer
device.
}

